[akg-scheduler] merge br_schedule into master

Created-by: ombre_mer1
Commit-by: ombre_mer1;MindSpore-Bot;huawuyi;corleoneliu;sasaki;TangZhiyun
Merged-by: corleoneliu
Description: <!--  Thanks for sending a pull request!  Here are some tips for you:

1) If this is your first time, please read our contributor guidelines: https://atomgit.com/mindspore/mindspore/blob/master/CONTRIBUTING.md

2) If you want to contribute your code but don't know who will review and merge, please add label `mindspore-assistant` to the pull request, we will find and do it as soon as possible.
-->

**What type of PR is this?**
<!-- 
选择下面一种标签替换下方 `/kind <label>`，可选标签类型有：
- /kind bug 
- /kind task
- /kind feature
- /kind refactor
- /kind clean_code
如PR描述不符合规范，修改PR描述后需要/check-pr重新检查PR规范。
-->
/kind feature

----

**What does this PR do / why do we need it**:
[akg-scheduler] merge br_schedule into master


----

**Which issue(s) this PR fixes**:
<!-- 
注意：在下方 #号 后仅输入issue编号或SR/AR编号，粘贴issue链接无效。
如PR描述不符合规范，修改PR描述后需要/check-pr重新检查PR规范。
-->
Fixes [#417](https://gitcode.com/mindspore/akg/issues/417) 

----

**Test Plan and Test result：What scenarios were tested, and what were the verification results（Function, performance, reliability, etc.）**：

----

**Self-checklist**:（**请自检，在[ ]内打上x，我们将检视你的完成情况，否则会导致pr无法合入**）

- [x] **设计**：PR对应的方案是否已经经过Maintainer评审，方案检视意见是否均已答复并完成方案修改
- [x] **测试**：PR中的代码是否已有UT/ST测试用例进行充分的覆盖，新增测试用例是否随本PR一并上库或已经上库
- [x] **验证**：PR描述信息中是否已包含对该PR对应的Feature、Refactor、Bugfix的预期目标达成情况的详细验证结果描述
- [ ] **接口**：是否涉及对外接口变更，相应变更已得到接口评审组织的通过，API对应的注释信息已经刷新正确
- [ ] **文档**：是否涉及官网文档修改，如果涉及请及时提交资料到Doc仓

<!-- **Special notes for your reviewers**: -->
<!-- - [ ] 是否导致无法前向兼容 -->
<!-- - [ ] 是否涉及依赖的三方库变更 -->


See merge request: mindspore/akg!2338

+++ akg-mlir/lib/Dialect/GPU/Transforms/KernelOutliningExt.cpp
@@ -128,7 +128,7 @@ static void initOperandOrder(mlir::func::FuncOp funcOp, SetVector<Value> &operan
       if (funcArguments[funcIdx] != operands[idx]) {
         continue;
       }
-      mapResult[idx] = funcIdx;
+      mapResult[idx] = static_cast<int>(funcIdx);
       break;
     }
   }
@@ -155,7 +155,7 @@ static void getAdditionalOperandOrder(mlir::func::FuncOp funcOp, SetVector<Value
       if (funcArguments[funcIdx] != rhs) {
         continue;
       }
-      additionalArgs[funcIdx] = idx;
+      additionalArgs[static_cast<int>(funcIdx)] = static_cast<int>(idx);
       break;
     }
   }
